refactor: MenuScreen 클래스 삭제에 따른 리펙토링
