+++ app/src/main/AndroidManifest.xml
@@ -22,17 +22,17 @@
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Swapit">
-
-        </activity>
-        <activity
-            android:name=".ui.auth.AuthActivity"
-            android:exported="true"
-            android:theme="@style/Theme.Swapit">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity
+            android:name=".ui.auth.AuthActivity"
+            android:exported="true"
+            android:theme="@style/Theme.Swapit">
+
         </activity>
 
     </application>
