+++ app/src/main/java/com/example/swap_it/MainActivity.kt
@@ -3,13 +3,17 @@ package com.example.swap_it
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import com.example.swap_it.ui.menu.MenuScreen
+import androidx.navigation.compose.rememberNavController
+
+import com.example.swap_it.ui.navigation.NavigationModule
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            MenuScreen().MenuScreen()
+            val navController = rememberNavController()
+            val navigationModule = NavigationModule()
+            navigationModule.NavigationGraph(navController)
         }
     }
 }
