+++ app/src/main/java/com/example/swap_it/ui/add_item/AddItemScreen.kt
@@ -13,8 +13,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextAlign
 import androidx.navigation.NavHostController
-import com.example.swap_it.ui.menu.MenuScreen
-import com.example.swap_it.ui.menu.MenuScreen.NavigationModule
+import com.example.swap_it.ui.component.AppBar
+import com.example.swap_it.ui.component.BottomNavigationBar
+
+import com.example.swap_it.ui.navigation.NavigationModule
 
 class AddItemScreen {
 
@@ -24,10 +26,10 @@ class AddItemScreen {
         val navigationModule = NavigationModule()
         Scaffold(
             topBar = {
-                MenuScreen().AppBar(navController = navController)
+                AppBar(navController = navController)
             },
             bottomBar = {
-                navigationModule.BottomNavigationBar(navController)
+                BottomNavigationBar(navController)
             },
 
             ) {
