+++ app/src/main/java/com/example/swap_it/ui/swap/ShoppingListScreen.kt
@@ -1,4 +1,4 @@
-package com.example.swap_it.ui.chat_list
+package com.example.swap_it.ui.swap
 
 import android.annotation.SuppressLint
 import androidx.compose.foundation.background
@@ -11,32 +11,36 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.Color.Companion.Gray
 import androidx.compose.ui.text.style.TextAlign
 import androidx.navigation.NavHostController
-import com.example.swap_it.ui.menu.MenuScreen
-import com.example.swap_it.ui.menu.MenuScreen.NavigationModule
+import com.example.swap_it.ui.component.AppBar
+import com.example.swap_it.ui.component.BottomNavigationBar
 
-class ChatListScreen {
-    @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter", "NotConstructor")
+
+import com.example.swap_it.ui.navigation.NavigationModule
+
+class ShoppingListScreen {
+    @SuppressLint("NotConstructor", "UnusedMaterial3ScaffoldPaddingParameter")
     @Composable
-    fun ChatListScreen(navController: NavHostController) {
+    fun ShoppingListScreen(navController: NavHostController) {
         val navigationModule = NavigationModule()
         Scaffold(
             topBar = {
-                MenuScreen().AppBar(navController = navController)
+                AppBar(navController = navController)
             },
             bottomBar = {
-                navigationModule.BottomNavigationBar(navController)
+                BottomNavigationBar(navController)
             },
 
-        ) {
+            ) {
             Box(
                 modifier = Modifier
                     .fillMaxSize()
-                    .background(Color.Magenta)
+                    .background(Gray)
             ) {
                 Text(
-                    text = "채팅",
+                    text = "거래",
                     style = MaterialTheme.typography.labelLarge,
                     textAlign = TextAlign.Center,
                     color = Color.White,
@@ -45,8 +49,6 @@ class ChatListScreen {
             }
         }
 
-
     }
-
 
 }
