+++ app/src/main/java/com/example/swap_it/ui/theme/Color.kt
@@ -15,4 +15,5 @@ val Gray3 = Color(0xFF767676)
 val Gray4 = Color(0xFFBDBDBD)
 val Gray5 = Color(0xFFE0E0E0)
 val Gray6 = Color(0xFFF3F3F3)
+val BackgroundColor = Color(0xFFF9F9F9)
 val White = Color(0xFFFFFFFF)
